refactor(velvet): consolidate Component test fixtures
Merge single-seam fixture files into topical fixtures (11 clusters:
ObjectIs predicates, widget pool helpers, fiber lifecycle, memo/ref
forwarding, VNode pool/builder, context re-render). Test bodies are
preserved verbatim; colocation is used where base classes or generic
parameterizations differ.
