--- Packages/com.velvet.core/Runtime/Component/Tests/Editor/AnimatePresenceListConvergenceTests.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 0127be9bfec4941819f69e2f5456af5f
