--- Packages/com.velvet.core/Runtime/Component/Tests/Editor/ComponentFiberLifecycleTests.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: ce3c50423fd054bb5b4a662dc1428276
