+++ Packages/com.velvet.core/Runtime/Component/Tests/Editor/ComponentMemoPropsBailTests.cs
@@ -19,6 +19,10 @@ namespace Velvet.Tests
     /// the child bails while the predicate returns true and re-renders when it returns false.</item>
     /// <item>A bailed child keeps its committed output (the new props are not adopted) until a re-render
     /// actually occurs.</item>
+    /// <item>The forwarded ref is part of a memoized forwardRef component's identity, not its props: when the
+    /// parent passes a DIFFERENT <see cref="Ref{T}"/> instance with equal props, the child must still re-run
+    /// <see cref="Hooks.UseImperativeHandle"/> so the new ref receives the handle — otherwise the props-bail
+    /// would swallow the ref change and the new ref would never be populated.</item>
     /// </list>
     /// </summary>
     /// <remarks>
@@ -43,6 +47,9 @@ namespace Velvet.Tests
             s_useAreEqual = false;
             s_compareOnlyId = false;
             s_usePlainChild = false;
+            s_refA = new Ref<IHandle>();
+            s_refB = new Ref<IHandle>();
+            s_setUseB = default;
         }
 
         [Test]
@@ -172,6 +179,23 @@ namespace Velvet.Tests
                 "areEqual(Id only) returns false when Id changes (a -> b), so the child re-renders");
         }
 
+        [Test]
+        public void Given_AMemoizedForwardRefChild_When_TheParentPassesADifferentRefWithEqualProps_Then_TheNewRefIsPopulated()
+        {
+            // Arrange — a memoized forwardRef child whose handle landed in ref A.
+            using var mounted = V.Mount(_root, V.Component(MemoRefParent, key: "memo-ref"));
+            Assume.That(s_refA.Current, Is.Not.Null, "Precondition: the child's handle populated ref A on mount");
+            Assume.That(s_refB.Current, Is.Null, "Precondition: ref B is empty before the swap");
+
+            // Act — the parent re-renders passing a DIFFERENT ref instance (ref B); props are unchanged (null).
+            _root.Q<Button>("swap-ref").SimulateClick();
+            mounted.FlushStateForTest();
+
+            // Assert — the ref change re-runs the imperative handle, so the new ref receives the handle.
+            Assert.That(s_refB.Current, Is.Not.Null,
+                "A changed forwarded ref must re-run UseImperativeHandle even when the memoized props bail.");
+        }
+
         #region Parent / child components
 
         private static int s_memoChildRenderCount;
@@ -217,6 +241,40 @@ namespace Velvet.Tests
             return s_usePlainChild
                 ? V.Component(PlainChildRender, props, key: "child")
                 : V.Component(MemoChildRender, props, key: "child");
+        }
+
+        #endregion
+
+        #region Memoized forwardRef identity
+
+        private interface IHandle { }
+
+        private sealed class Handle : IHandle { }
+
+        private static Ref<IHandle> s_refA;
+        private static Ref<IHandle> s_refB;
+        private static StateUpdater<bool> s_setUseB;
+
+        // Memoized so a parent re-render with equal (null) props bails its body — the ref change must override that.
+        [Component(Memoize = true)]
+        private static VNode MemoHandleChild()
+        {
+            var handleRef = Hooks.ForwardedRef<IHandle>();
+            Hooks.UseImperativeHandle(handleRef, () => new Handle(), System.Array.Empty<object>());
+            return V.Label(text: "child");
+        }
+
+        [Component]
+        private static VNode MemoRefParent()
+        {
+            var (useB, setUseB) = Hooks.UseState(false);
+            s_setUseB = setUseB;
+            var currentRef = useB ? s_refB : s_refA;
+            return V.Div(children: new VNode[]
+            {
+                V.Button(name: "swap-ref", onClick: () => setUseB.Invoke(_ => true)),
+                V.Component<IHandle>(MemoHandleChild, currentRef, key: "child"),
+            });
         }
 
         #endregion
