--- Packages/com.velvet.core/Runtime/Component/Tests/Editor/ContextLiveTrackingTests.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 64e13a9436da64b3a9eefd0496d90389
