--- Packages/com.velvet.core/Runtime/Component/Tests/Editor/ForwardedRefTests.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: e70402b62894b4a05a1bbaa71b3c6054
