--- Packages/com.velvet.core/Runtime/Component/Tests/Editor/LabelPoolHelperTests.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 1ba722c3221ad4e1a9a206625624eb25
