--- Packages/com.velvet.core/Runtime/Component/Tests/Editor/ListCompositionParityTests.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: fd92eee80cdb64758800ffb90e34a23e
