--- Packages/com.velvet.core/Runtime/Component/Tests/Editor/MemoizedForwardRefIdentityTests.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: e723592d660a04b7da2d0755fbb6c69e
