--- Packages/com.velvet.core/Runtime/Component/Tests/Editor/ObjectIsGenericStringTests.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 74f922c4a65cb47ad81fe10c9c8cec8f
