--- Packages/com.velvet.core/Runtime/Component/Tests/Editor/ObjectIsNullableValueTests.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 71fbed80c43c04bdab77451c11252e7e
