--- Packages/com.velvet.core/Runtime/Component/Tests/Editor/SliderPoolHelperTests.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 64d2c7ed82e3948c78b0561d7d5e1671
