--- Packages/com.velvet.core/Runtime/Component/Tests/Editor/TextFieldPoolHelperTests.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: b3fa2ac0d692f4e3b9792594d9de445e
