--- Packages/com.velvet.core/Runtime/Component/Tests/Editor/TogglePoolHelperTests.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: f342fd4e83a4343098c67b13fdccd9ab
