--- Packages/com.velvet.core/Runtime/Component/Tests/Editor/VMemoGeneratedOverloadTests.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: f8244277a035441d0aaa392e1dce518f
