--- Packages/com.velvet.core/Runtime/Component/Tests/Editor/VNodePoolOwnershipTests.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 84c7365bc60a64bf0899ccb96231503d
