+++ Packages/com.velvet.core/Runtime/Component/Tests/Editor/VNodePoolTests.cs
@@ -18,6 +18,12 @@ namespace Velvet.Tests
     /// cleared.</item>
     /// <item>The pool caps at its maximum size, so returning more instances than the cap forces new
     /// allocations on the next rents beyond the cap.</item>
+    /// <item>Ownership tracking mirrors the identity set the node-array pool already keeps: the V.* factories
+    /// accept caller-supplied props bags / event arrays that a component may cache and reuse across renders, so
+    /// returning an instance the pool did not itself rent out must not clear it or admit it into the shared
+    /// pool — only pool-owned instances may be cleared and recycled. A release scope stages returns so a bag
+    /// retired mid-pass cannot be rented back until the scope ends, preventing a second retirement of the same
+    /// bag within one pass from recycling a live renter's object.</item>
     /// </list>
     /// </summary>
     [TestFixture]
@@ -243,5 +249,179 @@ namespace Velvet.Tests
             Assert.That(newCount, Is.GreaterThanOrEqualTo(2),
                 "The pool caps its size, so rents beyond the cap require new allocations");
         }
+
+        #region Ownership tracking
+
+        [Test]
+        public void Given_UserPropsNotRentedFromPool_When_ReturnProps_Then_ItIsNotCleared()
+        {
+            // Arrange — a caller-owned props bag, as a component caching props across renders holds.
+            var props = new FiberElementProps { Tooltip = "Save" };
+
+            // Act
+            VNodePool.ReturnProps(props);
+
+            // Assert — the caller's instance is left untouched.
+            Assert.That(props.Tooltip, Is.EqualTo("Save"));
+        }
+
+        [Test]
+        public void Given_UserPropsWasPassedToReturn_When_AnotherRentFollows_Then_TheUserInstanceIsNotHandedOut()
+        {
+            // Arrange
+            var props = new FiberElementProps { Tooltip = "Save" };
+            VNodePool.ReturnProps(props);
+
+            // Act — an unrelated factory call rents from the pool.
+            var rented = VNodePool.RentProps();
+
+            // Assert — the caller-owned instance never entered the shared pool.
+            Assert.That(ReferenceEquals(rented, props), Is.False);
+        }
+
+        [Test]
+        public void Given_RentedProps_When_ReturnedAndRentedAgain_Then_ThePoolStillRecyclesIt()
+        {
+            // Arrange — a genuinely pool-owned props bag with a stale field.
+            var rented = VNodePool.RentProps();
+            rented.Tooltip = "stale";
+            VNodePool.ReturnProps(rented);
+
+            // Act
+            var again = VNodePool.RentProps();
+
+            // Assert — the pooled instance round-trips and comes back scrubbed.
+            Assert.That((ReferenceEquals(again, rented), again.Tooltip),
+                Is.EqualTo((true, (string)null)));
+        }
+
+        [Test]
+        public void Given_ARentedPropsReturnedTwice_When_TwoRentsFollow_Then_TheyAreDistinctInstances()
+        {
+            // Arrange — the recycle sweep can reach one bag through several retired trees (a baseline
+            // retired by its owner and again by a parent expansion), so a double return must recycle once.
+            var bag = VNodePool.RentProps();
+            VNodePool.ReturnProps(bag);
+            VNodePool.ReturnProps(bag);
+
+            // Act — two rents; a double-pooled bag would come back for both.
+            var first = VNodePool.RentProps();
+            var second = VNodePool.RentProps();
+
+            // Assert — the pool never handed the same instance out twice.
+            Assert.That(ReferenceEquals(first, second), Is.False);
+        }
+
+        [Test]
+        public void Given_ARentedNodeArrayReturnedTwice_When_TwoRentsFollow_Then_TheyAreDistinctInstances()
+        {
+            // Arrange — same double-return discipline for the children-array pool.
+            var array = VNodePool.RentNodeArray(3);
+            VNodePool.ReturnNodeArray(array);
+            VNodePool.ReturnNodeArray(array);
+
+            // Act
+            var first = VNodePool.RentNodeArray(3);
+            var second = VNodePool.RentNodeArray(3);
+
+            // Assert
+            Assert.That(ReferenceEquals(first, second), Is.False);
+        }
+
+        [Test]
+        public void Given_ARentedEventArrayReturnedTwice_When_TwoRentsFollow_Then_TheyAreDistinctInstances()
+        {
+            // Arrange — same double-return discipline for the single-event-array pool.
+            var array = VNodePool.RentSingleEventArray();
+            VNodePool.ReturnEventArray(array);
+            VNodePool.ReturnEventArray(array);
+
+            // Act
+            var first = VNodePool.RentSingleEventArray();
+            var second = VNodePool.RentSingleEventArray();
+
+            // Assert
+            Assert.That(ReferenceEquals(first, second), Is.False);
+        }
+
+        [Test]
+        public void Given_AnOpenReleaseScope_When_ABagIsReturnedInsideIt_Then_ARentCannotGetItBackUntilTheScopeEnds()
+        {
+            // Arrange — a reconcile-pass-shaped release scope with one bag returned mid-scope. A tree
+            // retired mid-pass must not be re-rentable by a later factory call in the SAME pass, or a
+            // second retired tree reaching the same bag would recycle the new renter's live object.
+            var bag = VNodePool.RentProps();
+            VNodePool.BeginReleaseScope();
+            try
+            {
+                VNodePool.ReturnProps(bag);
+
+                // Act — drain every pooled bag; the staged one must not be among them.
+                var sawStagedBag = false;
+                for (var i = 0; i < 16; i++)
+                {
+                    if (ReferenceEquals(VNodePool.RentProps(), bag)) sawStagedBag = true;
+                }
+
+                // Assert
+                Assert.That(sawStagedBag, Is.False,
+                    "A bag returned inside a release scope must stay un-rentable until the scope ends");
+            }
+            finally
+            {
+                VNodePool.EndReleaseScope();
+            }
+        }
+
+        [Test]
+        public void Given_AReleaseScopeThatStagedAReturn_When_TheScopeEnds_Then_TheBagIsPoolableAgain()
+        {
+            // Arrange — drain the pool first so the flush target has room and the flushed bag sits on
+            // top of the stack, then stage one return inside a scope.
+            for (var i = 0; i < 16; i++)
+            {
+                VNodePool.RentProps();
+            }
+            var bag = VNodePool.RentProps();
+            VNodePool.BeginReleaseScope();
+            VNodePool.ReturnProps(bag);
+            VNodePool.EndReleaseScope();
+
+            // Act — the flushed bag is on top of the pool stack.
+            var rented = VNodePool.RentProps();
+
+            // Assert
+            Assert.That(ReferenceEquals(rented, bag), Is.True,
+                "A staged return must reach the pool when the scope ends");
+        }
+
+        [Test]
+        public void Given_UserEventArrayNotRentedFromPool_When_ReturnEventArray_Then_ItIsNotCleared()
+        {
+            // Arrange — a caller-owned single-event array, as a component caching events would hold.
+            var events = new FiberEventBinding[] { new ClickedBinding { Handler = () => { } } };
+
+            // Act
+            VNodePool.ReturnEventArray(events);
+
+            // Assert — the caller's binding survives.
+            Assert.That(events[0], Is.Not.Null);
+        }
+
+        [Test]
+        public void Given_UserEventArrayWasPassedToReturn_When_AnotherRentFollows_Then_TheUserArrayIsNotHandedOut()
+        {
+            // Arrange
+            var events = new FiberEventBinding[] { new ClickedBinding { Handler = () => { } } };
+            VNodePool.ReturnEventArray(events);
+
+            // Act
+            var rented = VNodePool.RentSingleEventArray();
+
+            // Assert — the caller-owned array never entered the shared pool.
+            Assert.That(ReferenceEquals(rented, events), Is.False);
+        }
+
+        #endregion
     }
 }
