--- Packages/com.velvet.core/Runtime/Component/Tests/Editor/VParamsOverloadTests.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 144bf8fb74e5646af888950826facd15
